Update2 wallis.c

+++ wallis.c
@@ -27,7 +27,7 @@ int main(void) {
 float wallis_pi(int n) {
   float num = 1;
   for (int i=1; i<=n; i++) {
-    num = num * ((4*i*i)/(4*i*i - 1));
+    num = num * ((float)(4*i*i)/(4*i*i - 1));
   } 
   
   return 2 * num;
